Adding athentication route at /login and display all users in db at /
